requirements.txt
pypresence===4.2.0
psutil===5.8.0
python-dotenv==0.15.0
pywin32==300